requirements.txt
connected_components_3d==3.18.0
numpy==2.1.1
opencv_python==4.10.0.84
Pillow==10.4.0
PyQt5==5.15.11
pyqt5_plugins==5.15.9.2.3
PyQt5_sip==12.15.0
python-socketio==5.11.4
scipy==1.14.1
tifffile==2024.8.10
